+++ lib/bpm/errors.rb
@@ -18,6 +18,12 @@ module BPM
     end
   end
   
+  class InvalidPackageNameError < BPM::InvalidPackageError
+    def format_message(package)
+      "Package at #{package.root_path} name and directory do not match. (name: #{package.name}).  Change the directory or package.json name to match."
+    end
+  end
+  
   class PackageNotFoundError < BPM::Error
     def format_message(name, version)
       "Could not find eligible package for '#{name}' (#{version})"